% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MilitaryShiny.R
\name{shinyapp}
\alias{shinyapp}
\title{Shiny App}
\usage{
shinyapp()
}
\value{
The shiny app of Group-E
}
\description{
This will recall the Shiny-App created by Group-E regarding Performance Scoreboard for Recruitment
}
\author{
Group-E
}
